implement repositories
 - add respective jpa repositories
