+++ src/main/java/com/esteem/billingAndPayment/domain/CustomerCategory.java
@@ -3,6 +3,7 @@ package com.esteem.billingAndPayment.domain;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Id;
+import javax.persistence.ManyToOne;
 
 @Entity
 public class CustomerCategory extends Metadata {
@@ -12,6 +13,8 @@ public class CustomerCategory extends Metadata {
     private String name;
     @Column(columnDefinition = "TEXT", length = 1000)
     private String description;
+    @ManyToOne
+    private CustomerType type;
 
     public Long getId() {
         return id;
@@ -35,6 +38,14 @@ public class CustomerCategory extends Metadata {
 
     public void setDescription(String description) {
         this.description = description;
+    }
+
+    public CustomerType getType() {
+        return type;
+    }
+
+    public void setType(CustomerType type) {
+        this.type = type;
     }
 
 }
