+++ src/main/java/com/esteem/billingAndPayment/repo/AccountRepo.java
@@ -1,0 +1,9 @@
+package com.esteem.billingAndPayment.repo;
+
+import com.esteem.billingAndPayment.domain.Account;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AccountRepo extends JpaRepository<Account, Long> {
+
+}
