+++ src/main/java/com/esteem/billingAndPayment/repo/Charge_ProductRepo.java
@@ -1,0 +1,9 @@
+package com.esteem.billingAndPayment.repo;
+
+import com.esteem.billingAndPayment.domain.Charge_Product;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface Charge_ProductRepo extends JpaRepository<Charge_Product, Long> {
+
+}
