+++ src/main/java/com/esteem/billingAndPayment/repo/CustomerCategoryRepo.java
@@ -1,0 +1,7 @@
+package com.esteem.billingAndPayment.repo;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CustomerCategoryRepo extends JpaRepository<CustomerCategoryRepo, Long> {
+
+}
