+++ src/main/java/com/esteem/billingAndPayment/repo/ProductRepo.java
@@ -1,0 +1,9 @@
+package com.esteem.billingAndPayment.repo;
+
+import com.esteem.billingAndPayment.domain.Product;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProductRepo extends JpaRepository<Product, Long> {
+
+}
