+++ src/main/java/com/esteem/billingAndPayment/repo/SubscriptionService.java
@@ -1,0 +1,9 @@
+package com.esteem.billingAndPayment.repo;
+
+import com.esteem.billingAndPayment.domain.Subscription;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SubscriptionService extends JpaRepository<Subscription, Long> {
+
+}
